requirements.txt
requests>=2.31.0
pandas>=2.0.0
tenacity>=8.0.0
python-dotenv>=1.0.0
